backend/ai_interview/requirements.txt
# Django and Django Rest Framework
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.2.0

# Database (Choose based on your setup)
sqlite3  # If using SQLite (comes pre-installed with Python)

# AssemblyAI for Speech-to-Text
requests>=2.31.0  # Required for API calls to AssemblyAI

# Google Gemini API (For Question Generation)
google-generativeai>=0.3.1  

# NLP and Text Processing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0  # For cosine similarity in answer evaluation

# Authentication and Security
djangorestframework-simplejwt>=5.3.0  # JWT Authentication
django-environ>=0.11.2  # For environment variables

